Split monolith MCP into 5 servers + Affinity CRM + NEW_DEAL briefings + attachment pipeline
MCP server split:
- google-calendar-gmail → google (Calendar + Gmail)
- New: slack (channel search, message search, file download)
- New: lowercarbon (portfolio facts, documents via Vectorize)
- New: granola (meeting notes via OAuth PKCE + SSE)
- New: affinity (Deal Log, company search, notes via REST v2)

Attachment pipeline:
- download_attachment saves to per-company folder AND auto-sends to
  chat via IPC in one tool call (no separate send_file step needed)
- Per-company folder structure: companies/{slug}/attachments/ + notes.md
- send_file accepts file_path or base64 for follow-up sharing
- Host-side binary content filter drops leaked base64 from agent output
- Container→host path translation for /workspace/ipc/ bind mounts

NEW_DEAL briefing flow:
- Affinity Deal Log lookup (stage, team, source, pass reason)
- Forwarded email attachment handling (retry with recipient's msgId)
- Deal Notes labeled as AI-generated (from deck), not team-authored
- Deck downloaded + sent but NOT analyzed in initial briefing

Infrastructure:
- scripts/deploy-test.sh — full deploy cycle: --no-cache build,
  cached agent-runner sync, file/session cleanup, service restart
- Immediate ack for all requests including follow-ups
